Added Skills and added flow attachments

+++ api/db/tables/flows.table.ts
@@ -5,5 +5,5 @@ export const flowsTable = new Table('Flows', () => ({
   id: {type: DataType.INTEGER.UNSIGNED, primaryKey: true, autoIncrement: true, unique: true, allowNull: false},
   name: {type: DataType.STRING, allowNull: false, unique: true},
   description: {type: DataType.TEXT},
-  difficulty: {type: DataType.INTEGER.UNSIGNED},
+  difficulty: {type: DataType.TINYINT.UNSIGNED},
 }));
